handel delete previous object bug[1] also change pagination max to 3
[1]delete previous object bug:
	in the first run, the database is empty so the queriy that
	finds the previous object will return None and we face a
	"None object has no attribute called id" error, to handle
	I used an "if" to check if there is an object, then delete.

+++ home/views.py
@@ -5,7 +5,7 @@ from django.core.paginator import Paginator
 
 def home(request) -> None:
     objects_list = Sweat.objects.order_by("-created_at")
-    paginator = Paginator(objects_list, 4)
+    paginator = Paginator(objects_list, 3)
     page_num = request.GET.get('page')
     sweats = paginator.get_page(page_num)
     return render(request, 'home/home.html', {'sweats': sweats})
@@ -13,6 +13,7 @@ def home(request) -> None:
 
 def about(request) -> None:
     about = About.objects.last()
-    pre_about = About.objects.filter(id=about.id - 1)
-    pre_about.delete()
+    if about != None:
+        pre_about = About.objects.filter(id=about.id - 1)
+        pre_about.delete()
     return render(request, 'home/about.html', {'about': about})
